Most example files are not prefixed with example
The ones that are not are things like util/pageBreak.md and
util/col20.md. Anything in util can be condered to be unlikely to be
changed by a user.

Conversely, things prefixed with example are things that are considered
likely to be changed by a user, and therefore the example prefix is to
warm the user that modifying that file could result in data loss if they
re-run cvGenerateBase.